pyproject.toml
[tool.poetry]
name = "django-dynamic-tables"
version = "0.1.0"
description = ""
authors = ["David Lubomirov <davidlubomirov@gmail.com>"]
readme = "README.md"
packages = [{include = "django_dynamic_tables"}]

[tool.poetry.dependencies]
python = "3.8.*"
django = "^4.2.3"
djangorestframework = "^3.14.0"
django-filter = "^23.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 120